requirements.txt


requests

pandas

plotly


python-dotenv


# for testing:
pytest
